requirements.txt
setuptools==66.1.1
pip==23.0.1
nvidia-cusparselt-cu12==0.6.3
mpmath==1.3.0
typing_extensions==4.14.1
triton==3.3.1
sympy==1.14.0
nvidia-nvtx-cu12==12.6.77
nvidia-nvjitlink-cu12==12.6.85
nvidia-nccl-cu12==2.26.2
nvidia-curand-cu12==10.3.7.77
nvidia-cufile-cu12==1.11.1.6
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cublas-cu12==12.6.4.1
networkx==3.5
MarkupSafe==3.0.2
fsspec==2025.5.1
filelock==3.18.0
nvidia-cusparse-cu12==12.5.4.2
nvidia-cufft-cu12==11.3.0.4
nvidia-cudnn-cu12==9.5.1.17
Jinja2==3.1.6
nvidia-cusolver-cu12==11.7.1.2
torch==2.7.1
pytz==2025.2
urllib3==2.5.0
tzdata==2025.2
typing-inspection==0.4.1
threadpoolctl==3.6.0
smmap==5.0.2
six==1.17.0
PyYAML==6.0.2
pydantic_core==2.33.2
protobuf==6.31.1
platformdirs==4.3.8
packaging==25.0
joblib==1.5.1
idna==3.10
click==8.2.1
charset-normalizer==3.4.2
certifi==2025.7.9
annotated-types==0.7.0
sentry-sdk==2.32.0
scipy==1.16.0
requests==2.32.4
python-dateutil==2.9.0.post0
pydantic==2.11.7
lightning-utilities==0.14.3
gitdb==4.0.12
scikit-learn==1.7.0
pandas==2.3.1
GitPython==3.1.44
wandb==0.21.0
torchmetrics==1.7.4
pillow==11.3.0
numpy==2.2.6
opencv-python-headless
torchvision==0.22.1
tqdm==4.67.1
safetensors==0.5.3
hf-xet==1.1.5
huggingface-hub==0.33.4
timm==1.0.17
pyparsing==3.2.3
kiwisolver==1.4.8
fonttools==4.58.5
cycler==0.12.1
contourpy==1.3.2
matplotlib==3.10.3
einops==0.8.1
wadler_lindig==0.1.7
jaxtyping==0.3.2
Wand==0.6.13
tifffile==2025.6.11
termcolor==3.1.0
lazy_loader==0.4
imageio==2.37.0
scikit-image==0.23.2
fire==0.7.0
webencodings==0.5.1
wcwidth==0.2.13
pure_eval==0.2.3
ptyprocess==0.7.0
fastjsonschema==2.21.1
widgetsnbextension==4.0.14
websocket-client==1.8.0
webcolors==24.11.1
uri-template==1.3.0
types-python-dateutil==2.9.0.20250708
traitlets==5.14.3
tornado==6.5.1
tinycss2==1.4.0
soupsieve==2.7
sniffio==1.3.1
Send2Trash==1.8.3
rpds-py==0.26.0
rfc3986-validator==0.1.1
rfc3339-validator==0.1.4
pyzmq==27.0.0
python-json-logger==3.3.0
Pygments==2.19.2
pycparser==2.22
psutil==7.0.0
prompt_toolkit==3.0.51
prometheus_client==0.22.1
pexpect==4.9.0
parso==0.8.4
pandocfilters==1.5.1
overrides==7.7.0
nest-asyncio==1.6.0
mistune==3.1.3
lark==1.2.2
jupyterlab_widgets==3.0.15
jupyterlab_pygments==0.3.0
jsonpointer==3.0.0
json5==0.12.0
h11==0.16.0
fqdn==1.5.1
executing==2.2.0
defusedxml==0.7.1
decorator==5.2.1
debugpy==1.8.15
comm==0.2.3
bleach==6.2.0
babel==2.17.0
attrs==25.3.0
async-lru==2.0.5
asttokens==3.0.0
terminado==0.18.1
stack-data==0.6.3
rfc3987-syntax==1.1.0
referencing==0.36.2
matplotlib-inline==0.1.7
jupyter_core==5.8.1
jedi==0.19.2
ipython_pygments_lexers==1.1.1
httpcore==1.0.9
cffi==1.17.1
beautifulsoup4==4.13.4
arrow==1.3.0
anyio==4.9.0
jupyter_server_terminals==0.5.3
jupyter_client==8.6.3
jsonschema-specifications==2025.4.1
isoduration==20.11.0
ipython==9.4.0
httpx==0.28.1
argon2-cffi-bindings==25.1.0
jsonschema==4.25.0
ipywidgets==8.1.7
ipykernel==6.30.0
argon2-cffi==25.1.0
nbformat==5.10.4
jupyter-console==6.6.3
nbclient==0.10.2
jupyter-events==0.12.0
nbconvert==7.16.6
jupyter_server==2.16.0
notebook_shim==0.2.4
jupyterlab_server==2.27.3
jupyter-lsp==2.2.6
jupyterlab==4.4.5
notebook==7.4.4
jupyter==1.1.1
kornia_rs==0.1.9
kornia==0.8.1
